# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 모두취업 API
  description: AI 기반 채용 플랫폼의 프론트엔드-백엔드 협업을 위한 표준 API 규격서입니다.
  version: 1.0.0
servers:
  - url: https://api.pro-nlp.com/api
    description: Production Server
  - url: http://localhost:8080/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Recruit:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        startDate:
          type: string
          format: date
        deadline:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
    Portfolio:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [link, file, github, notion]
        description:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date
    CoverLetter:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        recruit_id:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /recruits:
    get:
      summary: 채용 공고 목록 조회
      tags: [Recruitments]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: category
          in: query
          schema: { type: string }
        - name: techStack
          in: query
          schema: { type: string }
        - name: keyword
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [latest, popular] }
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Recruit' }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      totalPages: { type: integer }
  /recruits/{id}:
    get:
      summary: 채용 공고 상세 조회
      tags: [Recruitments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Recruit'
                  - type: object
                    properties:
                      content: { type: string }
  /portfolios:
    get:
      summary: 내 포트폴리오 목록 조회
      tags: [Portfolios]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Portfolio' } }
                  meta: { type: object, properties: { total: { type: integer } } }
    post:
      summary: 새 포트폴리오 등록
      tags: [Portfolios]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Portfolio' }
      responses:
        '201':
          description: 생성 성공
...
  /cover-letters:
    get:
      summary: 자소서 목록 조회
      tags: [Cover Letters]
      security:
        - bearerAuth: []
      parameters:
        - name: recruit_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/CoverLetter' } }
    post:
      summary: 새 자소서 작성
      tags: [Cover Letters]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 생성 성공
...